Add the first part of autoconf/autmake/libtool support to distcache. Due to the use of "configure.ac", instead of the "configure.in" used by versions 2.13 and earlier, autoconf 2.5x is required.
So far, autoconf's findings aren't handled by any of the source files,
but building and installation work okay.

The check for OpenSSL needs a little improvement (see the comments). The
bulk of the macro defined in acinclude.m4 is to prevent libdistcache*
and libnal* from becoming dependent on libssl.so/libcrypto.so. This only
applies when building distcache as shared libraries something which, by
default, is disabled.

The "configure.in" hack in bootstrap.sh is a work-around needed to
persuade libtoolize to recognise that we're storing much of the
generated files in config/. My understanding (based on reading the code
in cvsweb) is that the problem will disappear with the next libtool
release. The work-around is probably preferable to having lots of
generated files scattered around the top-level directory.

.cvsignore files have been modified/created to reflect the new set of
generated files.