Now we live in GNU World, remove the IN_MAKEDEPEND hacks.
Also, allow test_session to build without SSL/TLS support if the headers
don't exist - this means that binary goo can be generated instead of real
SSL_SESSIONs to use as test data. In fact, it would be a good idea for this
to be command-line-selectable anyway, even when OpenSSL headers/libs are
available; it is more likely to uncover protocol handling problems in
server implementations that look at the session data it is managing. As
our builtin implementation doesn't care about the data it's
sending/receiving, it's not particularly relevant for us at this point.