This change is primarily for Solaris and other systems where functions such as socket() are located in other libraries.
The library checks in autoconf only say if the given function exists in
the given library. The function checks say if the given function exists
in one of the libraries checked beforehand or in one of the
implicitly-linked libraries (e.g. libc). Therefore, by moving the
library checks to before the function checks, "configure" will say, for
example on solaris, that libsocket.so exists and that it found the
socket() function, which will allow libnal/common.h to continue without
bailing out with an #error.