Patch from Joe Orton <jorton@redhat.com>
If the build system contains pkg-config, use that to establish compiler and
linker flags for building against OpenSSL.

Note (from Geoff): we may remove all tests for openssl anyway at some point
as we don't really need it, and removing it would greatly simplify aspects
of the build. However, as long as we are checking for openssl, this patch
makes sense.