Patch from Joe Orton <jorton@redhat.com>
Fixes include paths because config/config.h is generated in 'builddir' and
this can be different to 'srcdir'. Eg. this patch allows building in a
different directory to the distcache sources (eg. building as a user that
has read-only access to the distcache source tree) - the following;
    /path/to/distcache/configure --srcdir=/path/to/distcache
will construct Makefiles and compiled output in the current directory.