OpenSSL support is not required for now, but as the source that used it before is merely commented-out, do likewise with the autoconf stuff.
However in test/Makefile.am we should not need those hooks anyway so I'm
removing them. Having grappled hugely and horribly with openssl autoconf
issues in Apache and Stunnel of late, I have at least gained the small
payoff that I know of some better ways for us to get autoconf to probe for
and configure openssl headers and libraries.