Second step in some header re-organisation. This should correct compilation after my last commit broke it. I've also replaced my evolvingly ugly CHANGES note with something reasonably summaritive.
We should now always include libsys/pre.h as our first header, and include
libsys/post.h as our last header. Between the two, the principle to apply
is that API headers should always be included prior to internal-only
headers. This is because API headers should not depend on anything internal
- libsys/pre.h takes care of including any system headers and that should
be all that's required for using the API headers.