Add "-sockowner", "-sockgroup", and "-sockperms" command-line switches to dc_server to allow control over listeners created using unix domain sockets. If specified, these controls will be attempted on the filesystem path corresponding to the listening socket, but on typical systems this would only work when starting dc_server as root.
This is based on an idea and patch from Joe Orton - I've modified the
naming of the switches and added the group ownership control. I'm unaware
whether his original syntax is "in the wild" on redhat systems, but if it
is I may yet have to change back to his original naming scheme.